separated DOM manipulation from game logic
