+++ index.html
@@ -36,5 +36,6 @@
         </div>
     </main>
     <script src="index.js"></script>
+    <!-- <script src="logic.js"></script> -->
 </body>
 </html>
