+++ index.js
@@ -1,7 +1,10 @@
+/* eslint-disable import/prefer-default-export */
 /* eslint-disable no-return-assign */
 /* eslint-disable no-alert */
 
-let gameBoard = ['', '', '', '', '', '', '', '', ''];
+// eslint-disable-next-line prefer-const
+// eslint-disable-next-line import/no-mutable-exports
+const gameBoard = ['', '', '', '', '', '', '', '', ''];
 
 const playerConstructor = (name, token) => ({ name, token });
 
@@ -45,57 +48,9 @@ const gameBoardValue = (token, position) => {
   gameBoard[position] = token;
 };
 
-// eslint-disable-next-line
-function wins(board, who, counter) {
-  if (
-    (board[0] === board[1] && board[0] === board[2] && board[0] !== '')
-            || (board[0] === board[3] && board[0] === board[6] && board[0] !== '')
-            || (board[0] === board[4] && board[0] === board[8] && board[0] !== '')
-            || (board[3] === board[4] && board[3] === board[5] && board[3] !== '')
-            || (board[6] === board[7] && board[6] === board[8] && board[6] !== '')
-            || (board[1] === board[4] && board[1] === board[7] && board[1] !== '')
-            || (board[2] === board[5] && board[2] === board[8] && board[2] !== '')
-            || (board[2] === board[4] && board[2] === board[6] && board[2] !== '')
-  ) {
-    alert(`${who.name} ( ${who.token} ) wins the Game!!!`);
-    showBoard();
-    return true;
-  }
-  if (counter > 9) {
-    alert(' Its a tie :-( ');
-    showBoard();
-    return true;
-  }
-}
-
-const startGame = () => {
-  const play = document.querySelector('.game-board');
-  let turn = 1;
-  play.addEventListener('click', (e) => {
-    const cell = e.target;
-    const dataIndex = cell.getAttribute('data-index');
-    function mainFunc(playerAny, positionAny) {
-      cell.innerText = `${playerAny.token}`;
-      gameBoardValue(playerAny, positionAny);
-      turn += 1;
-      if (wins(gameBoard, playerAny, turn)) {
-        gameBoard = ['', '', '', '', '', '', '', '', ''];
-        turn = 1;
-        gameBoardReset();
-      }
-    }
-    if (cell) {
-      if (gameBoard[dataIndex] === '') {
-        if (turn % 2 !== 0) {
-          mainFunc(player1, dataIndex);
-        } else if (turn % 2 === 0) {
-          mainFunc(player2, dataIndex);
-        }
-      } else {
-        alert('position taken');
-      }
-    }
-  });
-};
-
-startGame();
+exports.showBoard = showBoard;
+exports.gameBoardValue = gameBoardValue;
+exports.gameBoardReset = gameBoardReset;
+exports.player1 = player1;
+exports.player2 = player2;
+exports.gameBoard = gameBoard;
